@@ -1,62 +1,63 @@
 module github.com/prometheus/node_exporter
 
-go 1.18
+go 1.19
 
 require (
+	github.com/alecthomas/kingpin/v2 v2.3.2
 	github.com/beevik/ntp v0.3.0
 	github.com/coreos/go-systemd/v22 v22.5.0
-	github.com/dennwc/btrfs v0.0.0-20221026161108-3097362dc072
-	github.com/ema/qdisc v0.0.0-20200603082823-62d0308e3e00
+	github.com/dennwc/btrfs v0.0.0-20230312211831-a1f570bd01a1
+	github.com/ema/qdisc v0.0.0-20230120214811-5b708f463de3
 	github.com/go-kit/log v0.2.1
 	github.com/godbus/dbus/v5 v5.1.0
 	github.com/hashicorp/go-envparse v0.1.0
-	github.com/hodgesds/perf-utils v0.5.1
+	github.com/hodgesds/perf-utils v0.7.0
 	github.com/illumos/go-kstat v0.0.0-20210513183136-173c9b0a9973
-	github.com/josharian/native v1.0.0
-	github.com/jsimonetti/rtnetlink v1.3.0
+	github.com/josharian/native v1.1.0
+	github.com/jsimonetti/rtnetlink v1.3.2
 	github.com/lufia/iostat v1.2.1
 	github.com/mattn/go-xmlrpc v0.0.3
-	github.com/mdlayher/ethtool v0.0.0-20220830195143-0e16326d06d1
-	github.com/mdlayher/netlink v1.7.0
+	github.com/mdlayher/ethtool v0.0.0-20221212131811-ba3b4bc2e02c
+	github.com/mdlayher/netlink v1.7.2
 	github.com/mdlayher/wifi v0.0.0-20220330172155-a44c70b6d3c8
-	github.com/opencontainers/selinux v1.10.2
-	github.com/prometheus/client_golang v1.14.0
-	github.com/prometheus/client_model v0.3.0
-	github.com/prometheus/common v0.37.0
-	github.com/prometheus/exporter-toolkit v0.8.2
-	github.com/prometheus/procfs v0.8.0
-	github.com/safchain/ethtool v0.2.0
-	github.com/soundcloud/go-runit v0.0.0-20150630195641-06ad41a06c4a
-	golang.org/x/sys v0.4.0
-	gopkg.in/alecthomas/kingpin.v2 v2.2.6
+	github.com/opencontainers/selinux v1.11.0
+	github.com/prometheus-community/go-runit v0.1.0
+	github.com/prometheus/client_golang v1.15.1
+	github.com/prometheus/client_model v0.4.0
+	github.com/prometheus/common v0.44.0
+	github.com/prometheus/exporter-toolkit v0.10.0
+	github.com/prometheus/procfs v0.10.0
+	github.com/safchain/ethtool v0.3.0
+	golang.org/x/exp v0.0.0-20230522175609-2e198f4a06a1
+	golang.org/x/sys v0.8.0
+	howett.net/plist v1.0.0
 )
 
 require (
-	github.com/alecthomas/template v0.0.0-20190718012654-fb15b899a751 // indirect
 	github.com/alecthomas/units v0.0.0-20211218093645-b94a6e3cc137 // indirect
 	github.com/beorn7/perks v1.0.1 // indirect
-	github.com/cespare/xxhash/v2 v2.1.2 // indirect
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/dennwc/ioctl v1.0.0 // indirect
 	github.com/go-logfmt/logfmt v0.5.1 // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
 	github.com/google/go-cmp v0.5.9 // indirect
 	github.com/jpillora/backoff v1.0.0 // indirect
-	github.com/kr/pretty v0.3.1 // indirect
-	github.com/matttproud/golang_protobuf_extensions v1.0.1 // indirect
-	github.com/mdlayher/genetlink v1.2.0 // indirect
-	github.com/mdlayher/socket v0.4.0 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/matttproud/golang_protobuf_extensions v1.0.4 // indirect
+	github.com/mdlayher/genetlink v1.3.1 // indirect
+	github.com/mdlayher/socket v0.4.1 // indirect
 	github.com/mwitkow/go-conntrack v0.0.0-20190716064945-2f068394615f // indirect
 	github.com/siebenmann/go-kstat v0.0.0-20210513183136-173c9b0a9973 // indirect
+	github.com/xhit/go-str2duration/v2 v2.1.0 // indirect
 	go.uber.org/atomic v1.7.0 // indirect
 	go.uber.org/multierr v1.6.0 // indirect
-	golang.org/x/crypto v0.0.0-20221012134737-56aed061732a // indirect
-	golang.org/x/net v0.5.0 // indirect
-	golang.org/x/oauth2 v0.0.0-20220909003341-f21342109be1 // indirect
-	golang.org/x/sync v0.1.0 // indirect
-	golang.org/x/text v0.6.0 // indirect
+	golang.org/x/crypto v0.8.0 // indirect
+	golang.org/x/net v0.10.0 // indirect
+	golang.org/x/oauth2 v0.8.0 // indirect
+	golang.org/x/sync v0.2.0 // indirect
+	golang.org/x/text v0.9.0 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/protobuf v1.28.1 // indirect
+	google.golang.org/protobuf v1.30.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
