@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build ((openbsd && !amd64) || darwin || dragonfly) && !nofilesystem
-// +build openbsd,!amd64 darwin dragonfly
+//go:build (darwin || dragonfly) && !nofilesystem
+// +build darwin dragonfly
 // +build !nofilesystem
 
 package collector
