@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build (darwin || freebsd || openbsd || linux) && !nouname
-// +build darwin freebsd openbsd linux
+//go:build (darwin || freebsd || openbsd || netbsd || linux) && !nouname
+// +build darwin freebsd openbsd netbsd linux
 // +build !nouname
 
 package collector
