@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build (darwin || freebsd || openbsd) && !nouname
-// +build darwin freebsd openbsd
+//go:build (darwin || freebsd || openbsd || netbsd) && !nouname
+// +build darwin freebsd openbsd netbsd
 // +build !nouname
 
 package collector
