@@ -29,9 +29,7 @@ import (
 import "C"
 
 type cpuFreqCollector struct {
-	cpuFreq    *prometheus.Desc
-	cpuFreqMax *prometheus.Desc
-	logger     log.Logger
+	logger log.Logger
 }
 
 func init() {
@@ -40,16 +38,6 @@ func init() {
 
 func NewCpuFreqCollector(logger log.Logger) (Collector, error) {
 	return &cpuFreqCollector{
-		cpuFreq: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, cpuCollectorSubsystem, "frequency_hertz"),
-			"Current CPU thread frequency in hertz.",
-			[]string{"cpu"}, nil,
-		),
-		cpuFreqMax: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, cpuCollectorSubsystem, "frequency_max_hertz"),
-			"Maximum CPU thread frequency in hertz.",
-			[]string{"cpu"}, nil,
-		),
 		logger: logger,
 	}, nil
 }
@@ -81,14 +69,14 @@ func (c *cpuFreqCollector) Update(ch chan<- prometheus.Metric) error {
 
 		lcpu := strconv.Itoa(cpu)
 		ch <- prometheus.MustNewConstMetric(
-			c.cpuFreq,
+			cpuFreqHertzDesc,
 			prometheus.GaugeValue,
 			float64(cpuFreqV.UintVal),
 			lcpu,
 		)
 		// Multiply by 1e+6 to convert MHz to Hz.
 		ch <- prometheus.MustNewConstMetric(
-			c.cpuFreqMax,
+			cpuFreqMaxDesc,
 			prometheus.GaugeValue,
 			float64(cpuFreqMaxV.IntVal)*1e+6,
 			lcpu,
