@@ -22,11 +22,11 @@ import (
 	"os"
 	"strconv"
 
+	"github.com/alecthomas/kingpin/v2"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/procfs/sysfs"
-	"gopkg.in/alecthomas/kingpin.v2"
 )
 
 const raplCollectorSubsystem = "rapl"
