@@ -391,7 +391,7 @@ func TestAbdstatsParsing(t *testing.T) {
 }
 
 func TestDbufstatsParsing(t *testing.T) {
-	dbufstatsFile, err := os.Open("fixtures/proc/spl/kstat/zfs/dbuf_stats")
+	dbufstatsFile, err := os.Open("fixtures/proc/spl/kstat/zfs/dbufstats")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -535,6 +535,15 @@ func TestPoolStateParsing(t *testing.T) {
 					t.Fatalf("Incorrect parsed value for degraded state")
 				}
 			}
+			if poolName == "pool2" {
+				if isActive != uint64(1) && stateName == "suspended" {
+					t.Fatalf("Incorrect parsed value for suspended state")
+				}
+				if isActive != uint64(0) && stateName != "suspended" {
+					t.Fatalf("Incorrect parsed value for suspended state")
+				}
+			}
+
 		})
 		file.Close()
 		if err != nil {
